Deeptect_webserver/requirements.txt
asgiref==3.3.4
click==7.1.2
Django==3.2
djangorestframework==3.12.4
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
paho-mqtt==1.5.1
pymata4==1.10
PyMySQL==1.0.2
pyserial==3.5
python-decouple==3.4
pytz==2021.1
rpi==0.0.1
RPi.GPIO-def==0.2
sqlparse==0.4.1
Werkzeug==1.0.1
